[modify] functionのビルドを調整
